+++ whatsnew.md
@@ -12,6 +12,7 @@
 ## Minor changes
 
 * [FO3/FNV] Flag 0 in BOOK means "Scroll"
+* [FO3/FNV] make "RuntimeScriptProfiler xSE Extension Log" available
 
 ## Shortened Parameters